requirements.txt
open-flamingo
more_itertools
python-dotenv
faiss-gpu
hydra-core
datasets
lightning
omegaconf
torch
tqdm
pycocotools
pycocoevalcap
tensorboard
transformers
